rename test (strategies)
