+++ tests/report/test_transformationStrategies.py
